requirements.txt
torch 
torchaudio 
transformers 
flask
google-generativeai
requests